appctype: move to types/appctype
Having a types package at the top level was almost certainly unintentional.

Updates: https://github.com/tailscale/corp/issues/15038
