+++ appc/appc.go
@@ -13,9 +13,9 @@ import (
 	"time"
 
 	"golang.org/x/net/dns/dnsmessage"
-	"tailscale.com/appctype"
 	"tailscale.com/metrics"
 	"tailscale.com/tailcfg"
+	"tailscale.com/types/appctype"
 	"tailscale.com/types/ipproto"
 	"tailscale.com/types/nettype"
 	"tailscale.com/util/clientmetric"
